% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIE.R
\name{GenData}
\alias{GenData}
\title{Generate data}
\usage{
GenData(n, rmatrix, mu = 0, sd = 1, empirical = FALSE)
}
\arguments{
\item{n}{number of the observations.}

\item{rmatrix}{correlation matrix}

\item{mu}{vector, theoretical mean of the generated data.}

\item{sd}{vector, the standard deviation}

\item{empirical}{logical. Whether the mean and covariance matrix refers
to the empirical or population. See mvrnorm() in MASS package.}
}
\description{
Simulated data using the correlation matrix
}
\examples{
m <- GenMatrix(3, rlist = list(c(1, 2, 0.5), c(1, 3, 0.3), c(2, 3, 0.7)))
D <- GenData(100, rmatrix = m)
head(D)

}
\seealso{
GenMatrix
}
\author{
Shuo Wang
}
