% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIE.R
\name{GenMatrix}
\alias{GenMatrix}
\title{Generate correlation matrix}
\usage{
GenMatrix(p, rlist = list(), varnames = NULL)
}
\arguments{
\item{p}{number of variables in the matrix}

\item{rlist}{a list to add values to matrix. The vectors in the list contain
three elements, with the first two being the position of names
of variables, and the last one being the correlations between them.}

\item{varnames}{variable names. If NULL, the varnames are defaulted as
x1, ..., xp.}
}
\description{
A shortcut to generate correlation matrix. The generated matrix is the
input of the GenData.
}
\examples{

m <- GenMatrix(3, rlist = list(c(1, 2, 0.5), c(1, 3, 0.3), c(2, 3, 0.7)))
m

m <- GenMatrix(3, rlist = list(c("x1", "x2", 0.5), c("x1", "x3", 0.3), 
                               c("x2", "x3", 0.7)))
m

}
\author{
Shuo Wang
}
