% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIE.R
\name{rie.c}
\alias{rie.c}
\title{Relative Information Error for continuous variables}
\usage{
rie.c(cor, rvars, svars, std = TRUE)
}
\arguments{
\item{cor}{covariance matrix}

\item{rvars}{relevant variables}

\item{svars}{selected variables}

\item{std}{logical. whether information should be standardized.}
}
\description{
Information error and relative information error for continuous variables following
the multinomial distribution using correlation matrix.
}
\examples{
m <- GenMatrix(3, rlist = list(c(1, 2, 0.5), c(1, 3, 0.3), c(2, 3, 0.7)))
rvars <- c("x1", "x2")
svars <- c("x2", "x3")
rie.c(m, rvars, svars)

}
\author{
Shuo Wang
}
